Fix TS-SRT output compile on MacOS

+++ src/output/output_tssrt.h
@@ -32,7 +32,7 @@ namespace Mist{
     bool pushOut;
     bool rawMode{false};
     uint64_t lastRawPacket{0};
-    uint64_t rawIdx{INVALID_TRACK_ID};
+    size_t rawIdx{INVALID_TRACK_ID};
     Util::ResizeablePointer packetBuffer;
     TS::Stream tsIn;
     TS::Assembler assembler;
